% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_connect.R
\name{redshift_connection}
\alias{redshift_connection}
\title{Connect to an FT Redshift cluster}
\usage{
redshift_connection(cluster_name, username, password)
}
\arguments{
\item{cluster_name}{The Redshift cluster you wish to connect to. Can be one
of "prod", "int" or "int2".}

\item{username}{Your Redshift username}

\item{password}{Your Redshift password}
}
\value{
A Redshift connection object.
}
\description{
Creates a Redshift connection object.
}
\examples{
\dontrun{
ft_redshift_connection <- redshift_connection(
  cluster_name = "prod",
  rstudioapi::askForPassword("Database user"),
  rstudioapi::askForPassword("Database password")
 )
}
}
