% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_query.R
\name{redshift_query}
\alias{redshift_query}
\title{Query an FT Redshift cluster.}
\usage{
redshift_query(sql_file, cluster, username, password)
}
\arguments{
\item{sql_file}{File name (and path) of a .sql you would like to run.}

\item{cluster}{The name of the cluster on which the query should run. Use
'int' for the INT cluster and 'analytics' for the Analytics cluster.}

\item{username}{Your Redshift username.}

\item{password}{Your Redshift passowrd.}
}
\value{
A dataframe of the query output.
}
\description{
\code{redshift_query} returns the output of a SQL query on one of the FT
Redshift clusters. You must be connected to the FT internal network to use
this function. If you are commiting code to  github, do not write your
redshift user name and password into the file. Instead create a .Renviron
file in your home directory and access the credentials via
\code{Sys.getEnv()} (see examples).
}
\examples{
\dontrun{
redshift_query("my_query.sql","int",Sys.getEnv("redshift_username"),Sys.getEnv("redshift_password"))
}

}
